+++ src/utils/styles.ts
@@ -10,11 +10,4 @@ export const layoutStyles = StyleSheet.create({
     flexDirection: 'row',
     alignItems: 'center',
   },
-  SingleModalContainer: {
-    paddingLeft: 0,
-    paddingRight: 0,
-    marginLeft: 0,
-    marginRight: 2,
-    marginTop: 4,
-  },
 });